started working on Server type

+++ osc.go
@@ -2,5 +2,5 @@ package gosc
 
 type NetAddr struct {
 	Addr string
-	Port uint8
+	Port int
 }
